+++ tests/testthat/test_bivariate_ggplot_call.R
@@ -3,7 +3,7 @@ context("bivariate_ggplot_call")
 test_that("bivariate_ggplot_call with numerics", {
   expect_match(
     bivariate_ggplot_call("numeric", "numeric") %>% deparse(width.cutoff = 300),
-    "geom_point\\(alpha = \\.alpha\\)"
+    "geom_point\\(alpha = \\.alpha\\, pch = 21)"
   )
   expect_match(
     bivariate_ggplot_call("numeric", "factor") %>% deparse(width.cutoff = 300),