+++ NEWS.md
@@ -5,6 +5,7 @@
 * Display variable labels in drop-down menu of data table (#393)
 * optional subsetting and ordering datasets for tm_data_table
 * use `teal.devel::chunks_push_data_merge` to include merge code into reproducibility code
+* more consistent coloring behavior between bivariate plots
 
 # teal.modules.general 0.2.3